# meet-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api for meet application.",
        "title": "meet API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/pack/prepare-upload-item": {
            "post": {
                "description": "Prepare the upload for an item of one pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Prepare Upload Pack Item",
                "parameters": [
                    {
                        "description": "Payload Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PrepareUploadPackItemDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_ResourceUploadUrlDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/publish": {
            "post": {
                "description": "Publish the pack ready for consumers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Publish Pack",
                "parameters": [
                    {
                        "description": "Payload Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PackDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/v1/pack/{modelNickName}": {
            "get": {
                "description": "Get all the packs from one model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Get Packs From Model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_PackDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/new": {
            "put": {
                "description": "Create a new pack for model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Create New Pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-dto_PackDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}": {
            "delete": {
                "description": "Delete one pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Delete Pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pack number",
                        "name": "packNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}/description": {
            "post": {
                "description": "Edit the pack description by the model or admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Edit Pack Description",
                "parameters": [
                    {
                        "description": "Payload Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EditPackDescriptionDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}/info": {
            "get": {
                "description": "Get information about the pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Get Pack Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pack number",
                        "name": "packNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_PackInfoDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}/items": {
            "get": {
                "description": "Get all active items from the pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Get Items From Pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pack number",
                        "name": "packNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_PackItemDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}/title": {
            "post": {
                "description": "Edit the pack title by the model or admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Edit Pack Title",
                "parameters": [
                    {
                        "description": "Payload Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EditPackTitleDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/pack/{modelNickName}/{packNumber}/{packItem}": {
            "delete": {
                "description": "Delete a item of one Pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Delete Pack Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pack number",
                        "name": "packNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "item number on the pack",
                        "name": "packItem",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.EditPackDescriptionDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 280
                }
            }
        },
...
        "handler.PackDto": {
            "type": "object",
            "properties": {
                "modelNickName": {
                    "type": "string"
                },
                "packNumber": {
                    "type": "integer"
                }
            }
        },
        "handler.PrepareUploadPackItemDto": {
            "type": "object",
            "properties": {
                "isPublic": {
                    "type": "boolean"
                },
                "modelNickName": {
                    "type": "string"
                },
                "packNumber": {
                    "type": "integer"
                },
                "typeCode": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_dto_PackDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PackDto"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.ApiResponse-array_dto_PackInfoDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PackInfoDto"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.ApiResponse-array_dto_PackItemDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PackItemDto"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.ApiResponse-array_dto_ResourceUploadUrlDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ResourceUploadUrlDto"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-dto_PackDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "$ref": "#/definitions/dto.PackDto"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-string": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}